🔖 1.9.4: ♻ Refactor `MemberBoard` into separate component
