+++ src/widgets/Layout.tsx
@@ -9,3 +9,10 @@ export const FVertical = styled.div`
   display: flex;
   flex-direction: column;
 `
+
+export const Board = styled.div`
+  position: absolute;
+  display: flex;
+  left: 43px;
+  top: 29px;
+`
